requirements.txt
discord.py>=2.0.0,<3
python-dotenv==1.0.0